CRUD Condominios

+++ Models/Condominium.cs
@@ -2,7 +2,7 @@
 {
     public class Condominium
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public int Id_Company { get; set; }
         public string? Name { get; set; }
         public string? Address { get; set; }
